Renamed slot_tests.rs to lot_tests.rs
